requirements.txt
streamlit==1.32.2
transformers==4.38.2
torch==2.2.1
accelerate==0.27.2
bitsandbytes==0.42.0
python-dotenv==1.1.0
sentencepiece==0.2.0 